+++ Business.Entities/UntaggedUPCBusinessModal.cs
@@ -6,7 +6,7 @@ namespace Business.Entities
     {  
         public int UntaggedUPCID { get; set; }
   
-        public int DescriptionID { get; set; }
+        public long DescriptionID { get; set; }
         public string Description { get; set; }
 
         public string UPCCode { get; set; }
